style(api): resolve remaining ruff lint (behavior-preserving) → ruff check . clean
Hand-fixes the 52 non-auto-fixable lint issues across 28 files so `ruff check .`
passes under ruff 0.16.x with zero suppressions. All behavior-preserving:
subprocess.run gets check=False (keeps no-return-code-check behavior), nested ifs
collapsed, unused unpacks →_, SIM103 negations rewritten explicitly, RUF012 real
constants annotated ClassVar, redundant exc args dropped from logger.exception,
implicit Optional made explicit. Backend suite: 1065 passed, 0 failed (baseline-identical).

+++ apps/api/routes/projects/animations.py
@@ -73,7 +73,7 @@ def _interpolate_params(from_state: dict, to_state: dict, t: float) -> dict:
             result[key] = from_val + (to_val - from_val) * t
             # Preserve int type if both sides were ints
             if isinstance(from_val, int) and isinstance(to_val, int):
-                result[key] = int(round(result[key]))
+                result[key] = round(result[key])
         else:
             # Non-numeric: snap halfway
             result[key] = to_val if t >= 0.5 else from_val
